requirements.txt
Flask==2.0.1
image==1.5.33
Jinja2==3.0.1
json-logging==0.0.1
Pillow==8.3.2
requests==2.26.0
Werkzeug==2.0.1
